Make location search icon into a general component
