+++ screens/LocationSearch/index.ts
@@ -3,4 +3,3 @@ export * from "./SearchResultsList"
 export * from "./SearchResultView"
 export * from "./models"
 export * from "./SearchBar"
-export * from "./Icon"
